admin manager

+++ app/src/main/java/com/example/onlyfanshop/model/BrandDTO.java
@@ -7,13 +7,13 @@ import java.io.Serializable;
 public class BrandDTO implements Serializable {
     // Backend trả "brandID" (int32) -> để an toàn có thể nhận cả "id"
     @SerializedName(value = "brandID", alternate = {"id"})
-    private Long brandID;
+    private Integer brandID;
 
     @SerializedName("name")
     private String name;
 
-    public Long getBrandID() { return brandID; }
-    public void setBrandID(Long brandID) { this.brandID = brandID; }
+    public Integer getBrandID() { return brandID; }
+    public void setBrandID(Integer brandID) { this.brandID = brandID; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
